Porting revision 441961 from 3.5 to trunk:
Because the ViewDockAreaPopup is a popup, it disappears when the mouse
is clicked outside of it. So when you click on the kmix icon in the
systray the popup disappears and then reappears again. I've added a
timer so that if another mouse event is received withing 0.3 seconds of
the first, it will be ignored. So now you can toggle the slider on and
off by clicking on the mixer icon, just like klipper.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=489680