kmix(pulse): Basic structure for getting information from pulse and registering for the relevant callbacks.
This adds a structure whereby the first pulse mixer object created
will connect to PA and register for the necessary callbacks. The
subsequent objects will not connect a second time.

What now needs to be done is to store the information about
each sink/source in some kind of QMap or similar such that
the information can be presented.

We currently employ an event loop that runs while we try and connect.
This loop is exited on failure or when we've received the callbacks
about sink/source that allow us to fill in our cache.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110850