kmix(pulse): Basic support for showing PA volumes.
This commit does not actually show the volumes yet, however it does read
the card info and sink/source info and generate an internal cache of this information.

As the kmix UI is geared around devices and not around playback and capture
we try and group up the cards to show them as such. Not all sinks/source
actually belong to cards so we have to virtualise this to a degree.

I'd rather create a similar approach to pavucontrol and have a single 'mixer'
in kmix for all Playback Devices and a separate 'mixer' for all
Capture Devices. A third 'mixer' would represent all individual
Applications. I will discuss this with Helio so we can decide the way forward.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110852