kmix(pulse): Implement a different sheme for display utilising three fixed tabs.
This more closely matches the concepts in pulseaudio and is easier to understand IMO.
The downside is that it does not match the typical 'one tab per device' approach
that users may be used to.

If this scheme is used, the way in which things are stored would be changed, so
the current display code is somewhat inefficient.

Also store the name and use that for the mixer device identifier such that preferences
will be saved more reliably. Keep a copy of the channel map too, which we'll
use to properly set volumes etc.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110853