kmix(pulse): Activate the group-by-card mode again.
It seems now that I look that MixDevices cannot handle arbitrary channel maps as I
previously thought, so we are going to have to create multiple MixDevices
to handle a single Pulse cvolume and somehow recombine it again.

I will probably use the name identifier for this in the future.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110856