kmix(pulse): Some Channel Map/Mask fixes. Now supports up to 5.1 sound control.
This commit maps the PA channel_map to the KMIX ChannelMask/ChannelID structure when
we receive info from the PA server. We can then use that to set and get volumes.
This now allows us to work with a standard 5.1 setup \o/

What does not yet work is hotplug and per-app controls. Stay tuned.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110861