kmix(pulse): Add support for per-application volume changing.
This is as yet incomplete due to not having yet solved the triggering of a repaint
due to device (and now app) hotplug. Once this is solved, this code
should slot into place and just work.
(As things stand it will only draw mixer sliders for applications
detected during startup - it will not remove apps no longer
present, nor will it draw new ones. Same is true for actual
devices too)

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110866