kmix(pulse): Ensure that the names of the MixDevices do not contain spaces
For sinks/sources this is generally true anyway, but for playback/capture
streams the 'name' is generally a nice application name which may
quite regularly contain spaces.

Also for playback, ensure the name contains the client name too
otherwise the wrong stream may be controlled. It may actually be more sensible
to use the index for the name in the case of applications
as at present if a client changes it's name, or the stream name changes
we wont be able to match it up to the MixDevice.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110867