kmix(pulse-related): Hopefully prevent a 'double KMixWindow' race.
This is a rather complicated race condition exposed as a result of PA integration.
The bug results from the fact that KMixWindow construction can now take
quite a while and actually loops it's own event loop for connecting to
(and potentially autospawning) the PA daeamon. This gives a very small
window of opportunity for two KMixWindows to be created. In order to
prevent this, I use a static flag rather than the m_kmix pointer
to determin if we've started or not. This, and suitible protection
against m_kmix not being set yet when we come to use it, should
prevent this from being a problem.

The bug report can be seen here:
https://qa.mandriva.com/show_bug.cgi?id=56893

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110897