kmix(pulse): Always use playback volumes to represent our widgets.
I was previously artificially using the captureVolume of the Volume for my capture streams
but this is actually not a good idea as it means I cannot show a mute button
for the capture sliders. The hasSwitch argument to the MixDevice will automatically show a
'Capture' selector for volumes with captureVolume, which we do not want.
We want a regular mute button, so just use playbackVolume for everything.
It simplifies things a bit too.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110898