kmix: When signalling that a mixer needs redrawn, just redraw what's needed, not the whole GUI.
This commit allows much finer grained redraws of a given tab. We intercept
the signal and do all the low level redrawing before passing it on to
higher levels for applying configs etc.

Currently the addSpacing() calls in ViewSliders::add() cause a leak of spacer
objects and as such the mixers migrate away from the edge over time. Will
fix that in the next commit.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110906