kmix: Change how the profiles are loaded.
When dealing with the profiles for PulseAudio mixers, I found that it was happily
loading the GUI Profile for 'Playback Devices' for the 'Capture Devices' tab
which resulted in it hiding all controls. While I'm not super familiar with the logic
here, I don't really see why all xml files in the user profile directory
are processed. In order to address this, I only load the XML files that start
with the mixer base name. This seems to be logical and works for the use cases
I've tested, but could conceivably cause some problems with other mixer backends.

The change ensures that files in the list are always *.xml files and thus
the further check for this can be removed.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110909