kmix: Reintroduce the 'always save profiles' bug which is needed.
This basically reverts 'kmix: Honour the saveConfig flag in recreateGUI()'
I've added extra comments about this so that it's not accidentally fixed again later.

I needed to remove this earlier for PA support but I no longer use
this method directly for redrawing the UI, so the bug no longer affects me.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110910