kmix: Ensure we always create MixDeviceWidgets for all MixDevices of the Mixer.
If our GUIProfile does not permit them to be displayed, we still want to
allow them to be configured in the 'Configure Channels...' menu entry.

Prior to this we would only display the MixDevices that are defined
in the profile. This generally does not affect e.g. ALSA as a GUIProfile
will generally limit what is physically shown to the user.

Hiding MixDevices is not stored in the GUI Profile (it's stored elsewhere)
so I have to question the creation of a Profile *at all* when
visiting the 'Configure Channels...' menu entry. Perhaps a
more appropriate fix is just to stop it doing that?

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110925