kmix: Rework the fallback GUIProfile in MixerToolBox
The current incarnation creates a single fallback profile, but in actual fact
this profile is a per-mixer incarnation.

This change creates a unique fallback profile for each mixer and provides
and accessor method so that we can compare current profile to fallback profile
which I'll use in an upcoming commit.

Note that as currently, there is no memory handling of this list. It
will grow for the lifetime of the application but as mixers tend not to
come and go, this shouldn't be a problem.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110926