kmix: Rework the setMixSet() method for the View* classes
This changes a pure virtual public method to a pure virtual protected one
with a regular wrapper of the original name.

There is no functionality change in this commit, it just allows me to do
some common setup before all setMixSet() calls as they currently stand.

This will become more useful in the next commit which will manipulate GUIProfiles
when a new dynamic device appears.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110927