kmix: Update the GUIProfile if a new control is added to a dynamic mixer
This change will automatically edit a GUIProfile if a new control
is added to the Mixer. This can happen quite easily with the PulseAudio backend
A previous approach was to always create the widgets but hide them,
however this wasn't really fully integrated into the GUIProfile concept.
With that approach, the GUIProfile itself would not be updated and
reordering the Controls would not work (the new controls would
not be added to the GUIProfile).

This approach, while somewhat convoluted, allows control orders to be
preserved and new controls to show up happily even if they
are not currently mentioned in the GUIProfile.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110928