kmix(pulse): Change the connection logic to avoid the use of a QEventLoop.
As discovered when using a QEventLoop with phonon, this can sometimes
lead to strange races and problems. Use the same technique here which is
to attempt the initial connection to PA in a blocking mode and, if
detected, disconnect and reattach with an integrated glib mainloop.

This also changes some variables to include an 's_' prefix where
none existed before, a check to make sure the glib mainloop is
available and some minor indentation fixes.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1110948