Fix for bug 220621, patch was taken from the bugreport. (branch 4.3).
Pleae note that for branch 4.3 the patch could not be tested (not even
compiled, due to source incompatibilites ("namespace Expiremental" and
"knotificationitem-1/knotificationitem.h"). As the patch is very important,
and proved to work in 2 other branches), I decided to commit immedatley.
I'll query for test support on bko, and relevant mailing lists.

I'll close this bug report once it has been applied to all relevant
branches
CCBUGS: 220621

svn path=/branches/KDE/4.3/kdemultimedia/kmix/; revision=1143057