kmix: Do not save/load mixer profiles for dynamic mixers (e.g. PulseAudio)
These mixer profiles are just 'odd' when it comes to PA, especially the streams (as opposed to devices).
When used under plain ALSA mixer profiles make sense to hide the unecessary complexity
of ALSA from most users, but PA already hides that complexity and building a filter
system on top of that makes little sense and leads to strange bugs/configurations.

CCBUG: 265317
CCBUG: 237239
CCBUG: 264835

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1226946