kmix: Generify an earlier fix to not show close/new tab buttons with PulseAudio.
This just removes the specific PulseAudio check and relies on the 'dynamic' mixers
In theory the same logic could apply to OSSv4 but I doubt anyone will bother.

Also remove the option to Launch pavucontrol. There is no check to see if it is
installed and all the functionality of pavucontrol is now available in kmix
or Speaker Setup which is already linked to in the menu.

CCBUG: 265317
CCBUG: 237239
CCBUG: 264835

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1226949