kmix: Fix the 'Event Sounds' slider.
When the volume/channel map parts of a stream restore rule are invalid
(which is allowed as these are optional in the restore rules) then we
cannot draw a slider. As we need to draw the slider for the event sounds
rule, silently fix it up.

Other restore rules need to be kept as is, for the Move option to work
correctly.

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1226951