gui: Fix dynamic mixer drawing.
There was a problem in a previous refactor (r1242598) which lead to a
condition where empty mixers (which is usual case for PulseAudio's
'Playback' and 'Capture' streams mixers) then none of the mixer
pointers would be added ot the _mixers collection. As we have a
specific mixer supplied, we simply ensure that we always include
this mixer too. This restores the drawing of the streams.

CCBUG: 281176

svn path=/trunk/KDE/kdemultimedia/kmix/; revision=1263755