Add FindCanberra.cmake to look for libcanberra.
Commit 1057f84c6587452d5f361f2c5f6529d8556fed1c has made kmix start
looking for libcanberra, however it relied on PKG_CHECK_MODULES()
directly for that.

In practice, this meant ${CANBERRA_LIBRARIES} was being set to
"-lcanberra" instead of "-l/path/to/libcanberra.so", which would have
happened if CMake's FIND_LIBRARY() call had been used instead. This
meant compilation was broken if libcanberra was not in a standard
linker path such as /usr/lib.

Fix that by writing a proper FindCanberra.cmake that only uses
pkg-config's output as a hint when trying to find libcanberra.

CCMAIL: esken@kde.org
CCMAIL: colin@mageia.org