pulse: Kill an event sound warning message.
Originally this was meant to explain to users why their event sounds were not showing up,
but users still ended up asking the question anyway so it doesn't really serve it's purpose
and just ends up spamming logs.

So downgrade it to a debug message but comment it out anyway.

CCBUGS: 319530