Volume up/down unmutes first, then does nothing else
In KDE4, the volume up and down keys only unmuted when the volume was
originally muted, without performing the extra volume change. I like
this behavior because it makes it very simple to restore the original
volume after a quick mute.

The former behavior where the volume was first unmuted, then brought
down to 0, and then increased a bit, caused an effective loss of the
original volume value. Not a big deal, but my muscle memory is already
trained.

REVIEW: 122866