Import FindGLIB2 and FindPulseAudio from ECM
Summary:
This seems backwards considering that both are in ECM now, but they
are source incompatible to the ones from KDELibs4Support.
We can't just depend on ECM 5.41.0 to solve this because the current
master of kmix is supposed to be released with KDE Applications 17.12
and 5.41.0 is only due for release on December 9th. Doing different
checks depending on the ECM version is also not an option because
ECM's version isn't increased in git yet, i.e. it's the same as for
the release of 5.40.0, which doesn't have the new find modules.
So this seems to be the best option with the downside that somebody
needs to remember to remove them when it's safe to depend on ECM
5.41.0.
This also allows us to get rid of KDELibs4 now.

Test Plan: Successfully buildt with ECM 5.40.0 and current master

Reviewers: #kde_applications, marten, lbeltrame, cgiboudeaux

Differential Revision: https://phabricator.kde.org/D8745