KDE Frameworks 5 support for hotplugging
As can be seen by the amount of old code commented out in KMixDeviceManager,
a rather simplistic approach is taken:  the Solid UDIs are only used to
identify a sound card and extract a device ID, using regexp matching.
The UDIs used elsewhere in the application are generated from the Solid
UDI.  As noted in the KF5 porting notes, OSS was never supported and
ALSA is assumed here.