Get rid of dependency on KInit
Summary:
KMix heavily depends on KDE-specific libraries and it's
hard to build it for other environments. At the same time
Plasma has a volume control applet, so KMix is of little
value to its users.

According to the KInit documentation, it gives a gain of ~150ms
on P-III 500 when starting an application, and on modern systems
the difference in startup times should be even less noticeable,
so remove this Tier 3 dependency.

Reviewers: marten

Reviewed By: marten

Differential Revision: https://phabricator.kde.org/D16161