Depend on KDE Frameworks 5.41
Summary:
Allows us to drop the copied FindGLIB2 and PulseAudio cmake modules.
5.41.0 was released in December 2017 and thus will be over a year
old when kmix 19.04 is planned to be released.
Also remove the distinction between minimal ECM and KF5 versions,
they are relesed with the same version since quite some time.
Furthermore raise the minimal Qt version to the 5.7.0, which is
what KF5 5.41.0 requires.

Differential Revision: https://phabricator.kde.org/D16971