Fix restoring of "Playback Streams" tab
Even if it is explicitly added by the user, it is not saved and
restored when KMix restarts.  The reason seems to be that no GUI profile
is saved for the tab because it is a dynamic mixer, but restoring
insists on having a profile available.

There seems to be support for fallback profiles in KMixWindow::recreateGUI(),
but bearing in mind the complexity of the code and the comments for the
we_need_a_fallback calculation, MPRIS2 profile fallback is handled as a
special case above.

This appears to work for ALSA, but should also be checked for PulseAudio
dynamic controls.