Configure Channels dialogue: Fix crash and tidy up
The crash happened if an item was moved from one list to another using
the arrow buttons, and then the same item was dragged.  The item move in
DialogViewConfiguration::moveSelection() used QListWidgetItem::clone()
to copy the item, but that was not implemented for the subclassed
DialogViewConfigurationItem;  not all of the item got copied and the
serializing failed.  There is no need to actually copy the item, so
it can just be moved from one list to another.

General code tidy up:
Eliminate public data members for DialogViewConfigurationItem
Use new connect() syntax
All three cases in DialogViewConfiguration::slotDropped() did the same thing
Rename _qlw -> _qlwActive for consistency