Use QtPaMainLoop instead of a full GLib event loop just for PulseAudio
Replaces Qt's GLib event loop with a simple wrapper that maps
PulseAudio timers, events etc. to the corresponding Qt stuff.

Does not stop the Glib library being loaded (because if Qt is built to
use it then it becomes a dependency of qtcore), but allows it to work
with QT_NO_GLIB set, removes the Glib dependency from KMix itself, and
integrates better with Qt in general (prioritization of events etc.).

Differential Revision: https://phabricator.kde.org/D28982