Fix naming and management of global shortcuts
Correctly separate the internal action name and the user readable text,
so that they show up with readable names in the Global Shortcuts KCM.

Only create volume global shortcuts if the control has adjustable
(playback or capture) volume.

Do not create any shortcuts for a virtual or dynamic control, and do not
show the shortcuts configuration dialogue if it would be empty.

Replace protected MixDeviceWidget data members with access functions.