Tidy up the MixDeviceWidget context menu generation
Let MixDeviceWidget::contextMenuEvent() handle the initial creation
and eventual showing of the menu.  Rename the pure virtual showContextMenu()
to createContextMenu() and have it only add the control-specific actions
to the menu.

Remove the "Hide" control action and MixDeviceWidget::setDisabled().
The action was only defined for a MDWEnum and was not accessible in the GUI.
The same effect can be achieved using "Configure Channels".