System tray: Open the main KMix window on "Restore"
It's not useful to pop up the volume control for this action, as there
is no position available and so the very small volume control popup
appears (and often doesn't get noticed by the user) at the top left of
the screen.  Do what would be expected from the action text and open
the main KMix window instead.