Ensure that the volume popup sliders and labels line up
The control alignment carried out for the main KMix window in
ViewSliders::configurationUpdate() was not being done for
the system tray popup, which meant that the controls may
not line up with long labels.  Move the alignment to
ViewBase::adjustControlsLayout() and do it for the system tray
popup also.

CCBUG:314280