Preferences dialogue: Show warnings and disable controls at creation time
This means the messages and controls that depend on whether PulseAudio
and Canberra are available.  The critera cannot change at runtime, and
setting them up statically instead of at the showEvent() avoids the odd
situation where opening the preferences dialogue from the popup volume
control (which opens it with the "Volume Control" page shown) causes
the "Volume overdrive is only available with PulseAudio" message widget
on the "General" page to be blank.