Remove ancient taglib leftover
I'm not sure what purpose this served, but from the history I could
easily go back to, it was introduced with a 12 years old commit titled
"Put in right place". Which makes me assume this is a leftover from
the monolithic kdemultimedia, which had a find_package(TAGLIB).
Anyway, Taglib is never searched for, nor is the variable set manually,
so include_directories is passed an empty value.