Unload and disable the KDED module on KMix startup
It's not clear exactly what the purpose of the KDED module is.
It seems to do exactly the same as KMix does itself, watching
for new hotplugged sound hardware and updating the internal data
structures and saved configration, but without a GUI.  It does
expose the mixers via DBus, but that access is available via the
same interface to the KMix application itself.

There are comments and hints in KMixApp that this was intended
for access by other applications, in particular KMilo which managed
laptop hotkeys back in the KDE3 days.  However, there are no known
applications that us it now.  The plasma-pulseaudio applet does not
use KMix or any of its interfaces.

If this change really does turn out to be benign then at some time
in the future the KDED module will be removed.