Move the global mixer list handling Mixer -> MixerToolBox
Because the Mixer class was an unholy mess of the actual handling
of one mixer via an instance of the class, together with the static
list of known mixers and access to it.  Made even more confusing by
allowing direct access to the modifiable list by any calling class.

The list is now handled in the MixerToolBox namespace and is internal
to it only, read only access is via a single access function.
This simplifies access compared with the former method where, for
example, MixerToolBox::possiblyAddMixer() called Mixer::mixers() to
get the list and then appended to it directly.