OSS: Handle hotplugging where ALSA is also active
Somewhat complicated, because in this case both ALSA and OSS may
deliver hotplug events.  So hotplug and unplug events are accumulated
and after a settling delay a decision is made as to whether to use
ALSA or OSS.  This can at least be reliably determined from the Solid
device, but there has to be some guessing where there is no existing
mixer.

Also handle correctly hotplug and unplug events, or vice versa,
happening in quick succession.