Detect backends via CMake instead of trusting platform definitions
Using a combination of looking for include files, installed
packages, and a test program to detect the OSS version.  Hopefully
a better approach than simply looking at the platform symbols in
kmix-backends.cpp, because it is not certain whether these are still
up to date or completely accurate today.

Compile all of the backends as separate source files, to eliminate
the confusing mixture of conditionals in CMakeLists.txt for some
backends and being included in source code for others.  Include
the backend data in MixerToolBox because it is only used there.