OSS: Fix volume feedback to work with this backend
Volume feedback was originally implemented in commit 5749abf2,
but at that time it was not clear how device names should be
specified to play the sound on the correct channel.  Names are
now correct for PulseAudio, ALSA and OSS, specified by the backend
and retained by the MixDevice as a "hardware ID" which is only used
for this purpose.

Removed a spurious ca_context_change_device() immediately after
the ca_context_play(), which reset the Canberra device to the
default immediately after playing the first sound.