Simplify hotplugging by using Solid UDIs throughout
But call them a "Hotplug ID" so as not to imply that they are specific
to Solid (although in this implementation they are unless someone
resurrects the Sun backend).

Ask the backends which support hotplugging whether a ID is recognised,
via MixerToolBox::acceptsHotplugid() instead of KMixDevice Manager
having to know about the format of ALSA and OSS UDIs.  Also make the
backend filtering work properly in MixerToolBox (there is no point in
making KMixDeviceManager to do the same) so that only hotplug events
on the preferred backend will be recognised.  This simplifies the logic
in KMixDeviceManager so that there is no need for the queuing and
prioritisation implemented by commit 823be92a.

In multi driver mode, hotplug events for all supported backends will
be recognised and acted on by the appropriate backend.  This is
hopefully what was intended.