Preferences dialogue: Retain the state of currently unplugged mixers
That is, the mixers which are shown in the system tray popup menu
and managed on the "Volume Control" tab.  More involved than may
be expected, because the enabled mixers are stored as a single
list which is rewritten every time the configuration is changed
with updated values from the GUI.  Mixers which are not present
are not shown in this list and hence would be removed from the
saved settings.

Now the mixers - even inactive ones - are saved indivdually with an
enable setting.  Mixers which are not present are shown in the GUI
but disabled, and are saved again with their previous settings.
This includes their icon name and readable name, to be used when
they cannot be obtained from an active mixer.

Removed the DialogChooseBackends class and source files, its name
was misleading (it is not a dialogue) and its interface was only
a simple list of mixers as above.

Unfortunately this shows a mess (a mix of PulseAudio categories
and physical devices) if running KMix successive times with and
without PulseAudio support.  This scenario is not likely.