PulseAudio: Fix showing 4 tabs when starting with a new configuration
The config file [Profiles] group could get into a state (or be
deliberately edited into a state) where a PulseAudio tab would not
show even though they are supposed to be fixed and there is no GUI
to add or restore them.  Allow them to be hidden deliberately if this
is really what is wanted, but properly show the tabs starting from a
new configuration or with the "--fallback" option.