Add command line option to run without the system tray volume control
Mainly useful for testing.  Selected for the current session
only via the new command line option "--nosystemtray", which does
not change the user preference setting.

Update the application startup and command line parsing to implement
the new option.  Add an enum type for the startup options, instead of
having to pass round yet another bool parameter in KMixApp and
KMixWindow.