Fix closing the main window when there is no system tray volume control
If the application needs to quit when the main window is closed,
it is not enough to rely on QGuiApplication::quitOnLastWindowClosed()
because the main window is not actually deleted (and it cannot be,
because it is held as a simple pointer in KMixApp).

Eliminate KMixWindow::hideOrClose() and KMixWindow::queryClose(),
and make all of the decisions in KMixWindow::closeEvent().  Do an
explicit QCoreApplication::quit() there if the application needs
to quit.

Simplify some of the tests in KMixWindow::recreateGUI().