OpenBSD sndio(7) backend support
This is an initial OpenBSD sndio(7) backend suport. It does not yet handle the mute interface and can still be extended.

However, with this version it is possible to control all input/output and application channels. Information from outside is also accepted and processed.

I have changed the CMake build system so that ONLY the sndio backend is used under OpenBSD. Everything else is useless under OpenBSD.